+++ src/extract_crate_docs.rs
@@ -89,7 +89,7 @@ fn parse_rustdoc_json(rustdoc_json: &str) -> Result<Crate, Report> {
         let actual = krate.format_version;
 
         let help = if actual > expected {
-            "update `cargo-insert-docs` or use `--toolchain nightly-2025-07-15`"
+            "update `cargo-insert-docs` or use `--toolchain nightly-2025-07-16`"
         } else {
             "upgrade your nightly toolchain"
         };