+++ src/main.rs
@@ -120,7 +120,7 @@ struct Args {
     /// updated to be compatible.
     ///
     /// With this argument you can choose a nightly version that is guaranteed to be compatible
-    /// with the current version of this tool, like `nightly-2025-07-15`.
+    /// with the current version of this tool, like `nightly-2025-07-16`.
     #[arg(long, default_value = "nightly", verbatim_doc_comment)]
     toolchain: String,